+++ wled00/xml.cpp
@@ -196,13 +196,13 @@ void sappends(char stype, const char* key, char* val)
 //get values for settings form in javascript
 void getSettingsJS(byte subPage, char* dest)
 {
-  //0: menu 1: wifi 2: leds 3: ui 4: sync 5: time 6: sec
+  //0: menu 1: wifi 2: leds 3: ui 4: sync 5: time 6: sec 7: DMX 8: usermods 9: sound
   DEBUG_PRINT(F("settings resp"));
   DEBUG_PRINTLN(subPage);
   obuf = dest;
   olen = 0;
 
-  if (subPage <1 || subPage >8) return;
+  if (subPage <1 || subPage >9) return;
 
   if (subPage == 1) {
     sappends('s',SET_F("CS"),clientSSID);
@@ -552,7 +552,7 @@ void getSettingsJS(byte subPage, char* dest)
     sappend('i',SET_F("CH13"),DMXFixtureMap[12]);
     sappend('i',SET_F("CH14"),DMXFixtureMap[13]);
     sappend('i',SET_F("CH15"),DMXFixtureMap[14]);
-    }
+  }
   #endif
 
   if (subPage == 8) //usermods
@@ -572,6 +572,6 @@ void getSettingsJS(byte subPage, char* dest)
     sappend('v',SET_F("LR"),i2swsPin);
     sappend('v',SET_F("CK"),i2sckPin);
     }
-    
+
   oappend(SET_F("}</script>"));
 }